unique_together fields instead of relying on computed fields
